Fixed an issue when saving the CSV after no data was loaded
If the user exited the program through the available options while the data from the CSV was not loaded, it replaced the CSV with an empty file. Now, if the user leaves without loading data previously, it will not replace the CSV file; instead, it will show the user that since data was not loaded, the CSV could not be saved.